CHANGE: reformat the project

+++ src/api/FileManager.h
@@ -4,7 +4,6 @@
 #include <iostream>
 #include <vector>
 #include <fstream>
-#include <algorithm>
 #include <string>
 #include "XMLParser.h"
 
@@ -13,45 +12,45 @@ using std::vector;
 
 class FileManager{
 	static vector<string> filePaths;
-	static int current;
+	static size_t current;
 public:
 	/**
 	 * @brief Add file to file manager
 	 * @param filePath Input filepath, must be wide string!
-	*/
+	 */
 	static bool addFile(const string &filePath);
 	/**
 	 * @brief Remove file by searching its name, return false if not exist
 	 * @param filePath Input filepath, must be wide string!
-	*/
+	 */
 	static bool removeFile(const string &filePath);
 	/**
 	 * @brief Remove file by using its index in the vector, return false if not exist or index is out of bound.
 	 * @param idx Index of the file in the vector
-	*/
+	 */
 	static bool removeFile(size_t idx);
 	static void clearFiles();
 	/**
 	 * @brief Validate if the file path leads to an accessible file.
 	 * @param filePath Input filepath, must be wide string!
-	*/
+	 */
 	static bool isFileExist(const string &filePath);
 	/**
 	 * @brief Get File Name from File Path, eg: /example/folder/Hi.txt -> Hi.txt
 	 * @param filePath Input filepath, must be wide string!
-	*/
+	 */
 	static string getFile(const string &filePath);
 	/**
 	 * @brief Get File Path in File Manager using index, return empty wide-string if not exist or index is out of bound.
 	 * @param idx Index of the file in the vector
-	*/
+	 */
 	static string getFile(size_t idx);
 	/** @brief Get a list of file's name */
 	static vector<string> getFileList();
 	/** @brief Set current file idx to target idx */
-	static void setCurrent(int idx);
+	static void setCurrent(size_t idx);
 	/** @brief Get current file idx */
-	static int getCurrent();
+	static size_t getCurrent();
 	/** @brief Get current file path */
 	static string getCurrentFile();
 };
